refactor(cli): extract shared + status modules; always open secrets vault
Continue the per-command module split and fix a feature-gating bug:

- commands/shared.rs — open_secrets, extract_vault_password, prompt_password
                       (vault/password helpers used across commands)
- commands/status.rs — StatusArgs + the `status` handler (run)

Fix: the secrets vault was only opened inside `#[cfg(feature = "tui")]`
blocks in the Setup and Onboard arms, so a non-TUI build never opened it.
The vault is now opened unconditionally in every build; only the interactive
wizard call (which lives in the TUI crate) remains feature-gated, with a
clear fallback message when TUI is absent.

Verified both `cargo check` (default) and `--no-default-features` compile
clean — the latter is the configuration the prior un-gating had broken.

cargo clippy --workspace --all-targets -- -D warnings: clean
cargo fmt --all -- --check: clean
